requirements.txt
# AI-Trend-Radar Requirements

# Core
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data Sources
arxiv>=2.1.0
google-generativeai>=0.3.0 # Optional for Gemini
openai>=1.0.0 # For DeepSeek via OpenAI-compatible API
beautifulsoup4>=4.12.0 # For scraping blogs if needed
feedparser>=6.0.0


# Analysis
pandas>=2.0.0
numpy>=1.24.0

# Interface
streamlit>=1.30.0
